+++ core-ng/src/main/java/core/framework/impl/web/bean/BeanClassNameValidator.java
@@ -13,7 +13,7 @@ import static core.framework.util.Strings.format;
 public class BeanClassNameValidator {
     final Map<String, Class<?>> beanClasses = Maps.newHashMap();
 
-    void validate(Class<?> beanClass) {
+    public void validate(Class<?> beanClass) {
         beanClasses.compute(Classes.className(beanClass), (key, previous) -> {
             if (previous != null && !previous.equals(beanClass))
                 throw new Error(format("found bean class with duplicate name which can be confusing, please use different class name, previousClass={}, class={}", previous.getCanonicalName(), beanClass.getCanonicalName()));