+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## Change log
 ### 6.12.2 (4/2/2019 - )
-* http: update undertow to 2.0.19
+* http: update undertow to 2.0.20
+* search: update es to 6.7.1
 * test: added assertEnumClass(class1).hasAllConstantsOf(class2)
+* http: update http().bean() to accept only one class (for simplification), support register enum
 
 ### 6.12.1 (4/1/2019)
 * kafka: fix kafka producer metrics can be NaN